+++ src/main/controllers/CopyToOtherPartyNotSystemUserController.ts
@@ -12,8 +12,8 @@ import { setUserCase } from './helpers/CaseHelpers';
 import { getCaptionTextForCopyToOtherParty, getRedirectPageUrlNotSystemUser } from './helpers/CopyToOtherPartyHelper';
 import { getCopyToOtherPartyError } from './helpers/ErrorHelpers';
 import { getPageContent } from './helpers/FormHelpers';
+import { getCancelLink } from './helpers/LinkHelpers';
 import { getLanguageParam } from './helpers/RouterHelpers';
-import { getCancelLink } from './helpers/Rule92NotSystemUserHelper';
 
 export default class CopyToOtherPartyNotSystemUserController {
   private readonly form: Form;
